+++ projects/fhi-angular-components/src/lib/fhi-tree-view-selection/fhi-tree-view-selection.component.ts
@@ -1,32 +1,39 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output, ViewEncapsulation } from '@angular/core';
 
-import { FhiTreeViewCheckboxItem as Item} from './fhi-tree-view-checkbox-item.model';
+import { FhiTreeViewSelectionItem as Item} from './fhi-tree-view-selection-item.model';
 
 @Component({
-  selector: 'fhi-tree-view-checkbox',
-  templateUrl: './fhi-tree-view-checkbox.component.html',
+  selector: 'fhi-tree-view-selection',
+  templateUrl: './fhi-tree-view-selection.component.html',
   encapsulation: ViewEncapsulation.None,
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class FhiTreeViewCheckboxComponent {
+export class FhiTreeViewSelectionComponent {
 
   @Input() enableCheckAll = false;
+  @Input() singleSelection = false;
   @Input() items: Item[] = [];
 
   @Output() itemsChange = new EventEmitter<Item[]>();
 
+  ngOnInit() {
+    if (this.enableCheckAll) {
+      this.singleSelection = false;
+    }
+  }
+
   ngOnChanges() {
     this.createIds(this.items, 1);
-    this.updateDesecendantState(this.items, true);
+    this.updateDecendantState(this.items, true);
   }
 
   toggleExpanded(item: Item) {
     item.isExpanded = !item.isExpanded;
   }
 
-  toggleChecked(id: number | string, multiToggle = false, checkeAll = false) {
-    this.updateCheckedState(id, this.items, multiToggle, checkeAll);
-    this.updateDesecendantState(this.items, false);
+  toggleChecked(id: number | string, multiToggle = false, checkAll = false) {
+    this.updateCheckedState(id, this.items, multiToggle, checkAll);
+    this.updateDecendantState(this.items, false);
     if (!multiToggle) {
       this.itemsChange.emit(this.items);
     }
@@ -50,23 +57,28 @@ export class FhiTreeViewCheckboxComponent {
     return items.every(item => item.isChecked)
   }
 
-  private updateCheckedState(id: number | string, items: Item[], multiToggle: boolean, checkeAll: boolean) {
+  private updateCheckedState(id: number | string, items: Item[], multiToggle: boolean, checkAll: boolean) {
     items.forEach(item => {
       if (item.id === id) {
         if (multiToggle) {
-          (checkeAll) ? item.isChecked = true : item.isChecked = false;
-        } else {
+          (checkAll) ? item.isChecked = true : item.isChecked = false;
+        } else if (!this.singleSelection) {
           item.isChecked = !item.isChecked;
+        } else if (this.singleSelection) {
+          item.isChecked = true;
         }
       }
+      if (this.singleSelection && item.id !== id) {
+        item.isChecked = null;
+      }
       if (item.children && item.children.length > 0) {
-        this.updateCheckedState(id, item.children, multiToggle, checkeAll);
+        this.updateCheckedState(id, item.children, multiToggle, checkAll);
       }
       item.descendantStateConfirmed = false;
     });
   }
 
-  private updateDesecendantState(items: Item[], initialState: boolean) {
+  private updateDecendantState(items: Item[], initialState: boolean) {
     items.forEach(item => {
       if (item.children && item.children.length > 0) {
         if (item.children.every(item => item.descendantStateConfirmed)) {
@@ -80,11 +92,11 @@ export class FhiTreeViewCheckboxComponent {
           }
           item.descendantStateConfirmed = true;
         }
-        this.updateDesecendantState(item.children, initialState);
+        this.updateDecendantState(item.children, initialState);
 
       } else if (!item.descendantStateConfirmed) {
         item.descendantStateConfirmed = true; // I.e. has no descendants
-        this.updateDesecendantState(this.items, initialState);
+        this.updateDecendantState(this.items, initialState);
       }
     });
   }