+++ projects/fhi-angular-components/src/public-api.ts
@@ -14,6 +14,6 @@ export * from './lib/fhi-tree-view-navigation/fhi-tree-view-navigation.component
 export * from './lib/fhi-tree-view-navigation/fhi-tree-view-navigation.module';
 export * from './lib/fhi-tree-view-navigation/fhi-tree-view-navigation-item.model';
 
-export * from './lib/fhi-tree-view-checkbox/fhi-tree-view-checkbox.component';
-export * from './lib/fhi-tree-view-checkbox/fhi-tree-view-checkbox.module';
-export * from './lib/fhi-tree-view-checkbox/fhi-tree-view-checkbox-item.model';
+export * from './lib/fhi-tree-view-selection/fhi-tree-view-selection.component';
+export * from './lib/fhi-tree-view-selection/fhi-tree-view-selection.module';
+export * from './lib/fhi-tree-view-selection/fhi-tree-view-selection-item.model';