+++ src/MOCK_DB_DATA/library-items/tree-view/_tree-view.data.ts
@@ -2,8 +2,10 @@ import { LibraryItem } from 'src/app/views/shared/models/library-item.model';
 
 import { TreeViewNavigation } from './tree-view-navigation';
 import { TreeViewCheckbox } from './tree-view-checkbox';
+import { TreeViewRadio } from './tree-view-radio';
 
 export const TreeViewData: LibraryItem[] = [
   ...TreeViewNavigation,
-  ...TreeViewCheckbox
+  ...TreeViewCheckbox,
+  ...TreeViewRadio
 ];