+++ src/MOCK_DB_DATA/library-items/tree-view/_tree-view.ids.ts
@@ -1,4 +1,5 @@
 export const TreeViewIds = {
   TreeViewNavigation: undefined,
-  TreeViewCheckbox: undefined
+  TreeViewCheckbox: undefined,
+  TreeViewRadio: undefined
 }